requirements.txt
Django==1.6.5
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.12
django-celery==3.1.10
kombu==3.0.20
pyserial==2.7
pytz==2014.4
wsgiref==0.1.2
